// src/enums/HttpStatus.enum.ts
export enum HTTPStatus {
	OK = 200,
	CREATED = 201,
	ACCEPTED = 202,
	BAD_REQUEST = 400,
	UNAUTHORIZED = 401,
	FORBIDDEN = 403,
	NOT_FOUND = 404,
	METHOD_NOT_ALLOWED = 405,
	CONFLICT = 409,
	CLIENT_CLOSED_REQUEST = 499,
	INTERNAL_SERVER_ERROR = 500,
	SERVICE_UNAVAILABLE = 503,
	GATEWAY_TIMEOUT = 504,
	FOUND = 302,
	MOVED_PERMANENTLY = 301,
	SEE_OTHER = 303,
	TEMPORARY_REDIRECT = 307,
	PERMANENT_REDIRECT = 308,
}